Small changes to database

+++ src/Arrangement.java
@@ -4,13 +4,13 @@ public class Arrangement {
 
     private int id;
     private String name;
-    private double totalPrice;
+    private float totalPrice;
     private int duration;
     private int numberOfEvents;
     private int participants;
     private ArrayList<Event> listOfEvents;
 
-    public Arrangement(int id, String name, double totalPrice, int duration, int numberOfEvents, int participations){
+    public Arrangement(int id, String name, float totalPrice, int duration, int numberOfEvents, int participations){
         listOfEvents = new ArrayList<>();
         this.id = id;
         this.name = name;
